% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_sim_crt.R
\name{gen_bin_crt}
\alias{gen_bin_crt}
\title{Simulate cluster-randomized trial data with binary outcomes}
\usage{
gen_bin_crt(
  m = c(1, 1),
  m_alloc = c(0.5, 0.5),
  n = c(1, 1),
  n_cv = c(0, 0),
  p_vec = c(0.5, 0.5),
  rho = c(0, 0)
)
}
\arguments{
\item{m}{Number of clusters.}

\item{m_alloc}{Allocation ratio of clusters per arm. Default is
\code{c(0.5, 0.5)}.}

\item{n}{Mean size of each cluster.}

\item{p_vec}{Probabilities of event for control and treatment groups.}

\item{rho}{Intraclass correlation coefficient. Default value is 0.}
}
\value{
Simulated binary outcomes represented as an n x 4 matrix with the
 columns encoding the treatment arm, cluster, individual, and response.
}
\description{
Simulate cluster-randomized trial data with binary outcomes
}
\author{
Lee Ding \email{lee_ding@g.harvard.edu}
}
