% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_design_crt.R
\name{gs_design_crt}
\alias{gs_design_crt}
\title{Compute stopping boundaries, maximum sample size, and expected sample
 sizes for a group sequential cluster randomized trial.}
\usage{
gs_design_crt(
  k = 3,
  outcome_type = 1,
  test_type = 1,
  test_sides = 1,
  size_type = 1,
  timing_type = 2,
  recruit_type = 1,
  delta = 1,
  sigma_vec = c(1, 1),
  p_vec = c(0.5, 0.5),
  rho = 0,
  alpha = 0.05,
  beta = 0.1,
  m = 2,
  m_alloc = c(0.5, 0.5),
  n = 1,
  n_cv = c(0, 0),
  info_timing = 1,
  m_timing = c(1, 1),
  n_timing = c(1, 1),
  alpha_sf = sf_ldof,
  alpha_sf_par = -4,
  beta_sf = sf_ldof,
  beta_sf_par = -4,
  tol = 1e-06,
  r = 18
)
}
\arguments{
\item{k}{Number of analyses planned, including interim and final.}

\item{outcome_type}{\code{1=}continuous difference of means
\cr \code{2=}binary difference of proportions}

\item{test_type}{\code{1=} early stopping for efficacy only
\cr \code{2=} early stopping for binding futility only
\cr \code{3=} early stopping for non-binding futility only
\cr \code{4=} early stopping for either efficacy or binding futility
\cr \code{5=} early stopping for either efficacy or non-binding futility}

\item{test_sides}{\code{1=} one-sided test \cr \code{2=} two-sided test}

\item{size_type}{\code{1=}clusters per arm \cr \code{2=}cluster size}

\item{timing_type}{\code{1=} maximum and expected sample sizes based on
specified information fractions in \code{info_timing}; recruit according to
design specified in \code{recruit_type}
\cr \code{2=} maximum sample size based on specified information fractions
in \code{info_timing} and expected sample sizes based on specified sample
size fractions in \code{m_timing} and \code{n_timing}
\cr \code{3=} maximum and expected sample sizes based on specified
sample size fractions in \code{m_timing} and \code{n_timing}
@param recruit_type \code{1=}recruit clusters with fixed sizes
\cr \code{2=}recruit individuals into fixed number of clusters
\cr \code{3=}recruit at both cluster and individual levels}

\item{delta}{Effect size for theta under alternative hypothesis. Must be > 0.}

\item{sigma_vec}{Standard deviations for control and treatment groups
(continuous case).}

\item{p_vec}{Probabilities of event for control and treatment groups
(binary case).}

\item{rho}{Intraclass correlation coefficient. Default value is 0.}

\item{alpha}{Type I error, default value is 0.05.}

\item{beta}{Type II error, default value is 0.1 (90\% power).}

\item{m}{Number of clusters for finding maximum mean cluster size. If
\code{m} is a scalar, it is treated as the total number of clusters across
arms. If \code{m} is a vector of length 2, it is treated as the number of
clusters per arm.}

\item{m_alloc}{Allocation ratio of clusters per arm. Default is
\code{c(0.5, 0.5)}.}

\item{n}{Mean cluster size for finding maximum number of clusters. If
\code{n} is a scalar, it is treated as the average cluster size for both
arms. If \code{n} is a vector of length 2, it is treated as the average
cluster size per arm.}

\item{n_cv}{Coefficient of variation for cluster size. If \code{n_cv} is a
scalar, it is treated as the coefficient of variation for both arms. If
\code{n_cv} is a vector of length 2, it is treated as the coefficient of
variation per arm.}

\item{info_timing}{Sets timing of interim analyses based on information
fractions. Default of 1 produces analyses at equal-spaced increments.
Otherwise, this is a vector of length \code{k} or \code{k-1}. The values
should satisfy \code{0 < info_timing[1] < info_timing[2] < ... <
info_timing[k-1] < info_timing[k]=1}.}

\item{m_timing}{Sets timing of interim analyses based on fractions of the
number of clusters.}

\item{n_timing}{Sets timing of interim analyses based on cluster size}

\item{alpha_sf}{A spending function or a character string indicating an
upper boundary type (that is, \dQuote{WT} for Wang-Tsiatis bounds,
\dQuote{OF} for O'Brien-Fleming bounds, and \dQuote{Pocock} for Pocock
bounds). The default value is \code{sf_ldof} which is a Lan-DeMets
O'Brien-Fleming spending function. See details,
\code{vignette("spending_function_overview")}, manual and examples.}

\item{alpha_sf_par}{Real value, default is \eqn{-4} which is an
O'Brien-Fleming-like conservative bound when used with a
Hwang-Shih-DeCani spending function. This is a real-vector for many spending
functions. The parameter \code{alpha_sf_par} specifies any parameters needed
for the spending function specified by \code{alpha_sf}; this will be ignored
for spending functions (\code{sf_ldof}, \code{sf_ld_pocock}) or bound types
(\dQuote{OF}, \dQuote{Pocock}) that do not require parameters.}

\item{beta_sf}{A spending function or a character string indicating an
lower boundary type (that is, \dQuote{WT} for Wang-Tsiatis bounds,
\dQuote{OF} for O'Brien-Fleming bounds, and \dQuote{Pocock} for Pocock
bounds). The default value is \code{sf_ldof} which is a Lan-DeMets
O'Brien-Fleming spending function. See details,
\code{vignette("spending_function_overview")}, manual and examples.}

\item{beta_sf_par}{Real value, default is \eqn{-4} which is an
O'Brien-Fleming-like conservative bound when used with a
Hwang-Shih-DeCani spending function. This is a real-vector for many spending
functions. The parameter \code{beta_sf_par} specifies any parameters needed
for the spending function specified by \code{beta_sf}; this will be ignored
for spending functions (\code{sf_ldof}, \code{sf_ld_pocock}) or bound types
(\dQuote{OF}, \dQuote{Pocock}) that do not require parameters.}

\item{tol}{Tolerance for error (default is 0.000001). Normally this will not
be changed by the user.  This does not translate directly to number of
digits of accuracy, so use extra decimal places.}

\item{r}{Integer value controlling grid for numerical integration as in
Jennison and Turnbull (2000); default is 18, range is 1 to 80.  Larger
values provide larger number of grid points and greater accuracy.  Normally
\code{r} will not be changed by the user.}
}
\value{
Object containing the following elements: \item{k}{As
 input.} \item{outcome_type}{As input.} \item{test_type}{As input.}
 \item{test_sides}{As input.} \item{size_type}{As input.}
 \item{recruit_type}{As input.} \item{timing_type}{As input.} \item{delta}{As
 input.} \item{sigma_vec}{As input.} \item{p_vec}{As input.} \item{rho}{As
 input.} \item{alpha}{As input.} \item{beta}{As input.} \item{info_timing}{As
 input.} \item{m_timing}{As input.} \item{n_timing}{As input.}
 \item{info_schedule}{Fraction of maximum information at each planned interim
 analysis based on \code{timing_type} and \code{info_timing}.}
 \item{m_schedule}{Fraction of maximum number of clusters per arm at each
 planned interim analysis based on \code{timing_type} and \code{m_timing}.}
 \item{n_schedule}{Fraction of maximum cluster size at each planned interim
 analysis based on \code{timing_type} and \code{n_timing}.} \item{i}{Fisher
 information at each planned interim analysis based on \code{timing_type}.}
 \item{max_i}{Maximum information corresponding to design specifications.}
 \item{m}{Number of clusters per arm at each planned interim analysis.}
 \item{max_m}{Maximum number of clusters per arm.} \item{e_m}{A vector of
 length 2 with expected number of clusters per arm under the null and
 alternative hypotheses. For simplicity, the expected sizes with non-binding
 futility boundaries are calculated assuming the boundaries are binding
 futility.} \item{n}{Mean cluster size at each planned interim analysis.}
 \item{max_n}{Maximum mean cluster size.} \item{e_n}{A vector of length 2
 with expected mean cluster sizes under the null and alternative hypotheses.
 For simplicity, the expected sizes with non-binding futility boundaries are
 calculated assuming the futility boundaries are binding.}
 \item{max_total}{Maximum number of individuals in the trial.}
 \item{e_total}{A vector of length 2 with expected total number of
 individuals in the trial under the null and alternative hypotheses. For
 simplicity, the expected sizes with non-binding futility boundaries are
 calculated assuming the futility boundaries are binding.} \item{sufficient}{
 Value denoting whether calculated sample size will be sufficient to achieve
 specified Type I error rate and power given the trial specifications.}
 \item{lower_bound}{Calculated lower futility boundaries under analysis
 schedule specified by \code{timing_type}} \item{upper_bound}{Calculated
 upper efficacy boundaries under analysis schedule specified by
 \code{timing_types}.} \item{tol}{As input.} \item{r}{As input.}
}
\description{
\code{gs_design_crt()} is used to determine the maximum sample
 size needed for a specified parallel group sequential cluster randomized
 trial to detect a clinically meaningful effect size with some Type I error
 rate and power. Code adapted from gsDesign package.
}
\references{
Jennison C and Turnbull BW (2000), \emph{Group Sequential
 Methods with Applications to Clinical Trials}. Boca Raton: Chapman and Hall.

Anderson K (2025). gsDesign: Group Sequential Design.
 R package version 3.8.0, https://keaven.github.io/gsDesign/.
}
\author{
Lee Ding \email{lee_ding@g.harvard.edu}
}
