% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsBoundCRT.R
\name{gs_lower_crt}
\alias{gs_lower_crt}
\title{Boundary derivation for binding or non-binding futility stopping only.}
\usage{
gs_lower_crt(
  theta = 0,
  I,
  false_neg,
  b = NULL,
  sides = 1,
  binding = TRUE,
  tol = 1e-06,
  r = 18,
  printerr = 0
)
}
\arguments{
\item{theta}{Effect size under null hypothesis. Default value is 0.}

\item{I}{Information levels at interim analyses for which stopping
boundaries are calculated. At any given interim analysis this should include
the information levels at the current and previous analyses.}

\item{false_neg}{Type II error spent for the interim analyses at the
information levels specified by I.}

\item{b}{Upper efficacy boundary values for the interim analyses at the
information levels specified by I.}

\item{sides}{\code{1=} 1-sided test (default) \cr \code{2=} 2-sided test}

\item{binding}{\code{TRUE=} binding (default) \cr \code{FALSE=} non-binding}

\item{tol}{Tolerance for error (default is 0.000001). Normally this will not
be changed by the user.  This does not translate directly to number of
digits of accuracy, so use extra decimal places.}

\item{r}{Integer value controlling grid for numerical integration as in
Jennison and Turnbull (2000); default is 18, range is 1 to 80.  Larger
values provide larger number of grid points and greater accuracy.  Normally
\code{r} will not be changed by the user.}

\item{printerr}{Print output for debugging.}
}
\value{
Object containing the following elements: \item{k}{Number of interim
 analyses.} \item{theta}{As input.} \item{I}{As input.} \item{a}{Computed
 lower futility boundaries at the specified interim analyses.} \item{b}{As
 input.} \item{r}{As input.} \item{error}{error flag returned; 0 if
 convergence; 1 indicates error.}
}
\description{
\code{gs_lower_crt} is used to calculate the stopping boundaries
 for a group sequential trial with 1 or 2-sided futility-only stopping. Code
 adapted from gsDesign package.
}
\references{
Jennison C and Turnbull BW (2000), \emph{Group Sequential
 Methods with Applications to Clinical Trials}. Boca Raton: Chapman and Hall.
}
\author{
Lee Ding \email{lee_ding@g.harvard.edu}
}
