% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_design_crt.R
\name{gs_plot_crt}
\alias{gs_plot_crt}
\title{Plot group sequential CRT design}
\usage{
gs_plot_crt(x)
}
\arguments{
\item{x}{Object of class \code{gs_design_crt} created by \code{gs_design_crt()}.}
}
\value{
Plot of stopping boundaries for the group sequential CRT design
}
\description{
Plot group sequential CRT design
}
\author{
Lee Ding \email{lee_ding@g.harvard.edu}
}
