% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_design_crt.R
\name{gs_print_crt}
\alias{gs_print_crt}
\title{Print group sequential CRT design}
\usage{
gs_print_crt(x)
}
\arguments{
\item{x}{Object of class \code{gs_design_crt} created by \code{gs_design_crt()}.}
}
\value{
Printed output of design parameters and table of stopping boundaries
}
\description{
Print group sequential CRT design
}
\author{
Lee Ding \email{lee_ding@g.harvard.edu}
}
