% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsBoundCRT.R
\name{gs_probability_crt}
\alias{gs_probability_crt}
\title{Compute stopping boundary crossing probabilities.}
\usage{
gs_probability_crt(theta = 0, I = 1, a = 0, b = 1, sides = 1, r = 18)
}
\arguments{
\item{theta}{Effect size. Default value is 0.}

\item{I}{Information levels at interim analyses for which boundary crossing
probabilities are computed.}

\item{a}{Lower futility boundaries for the interim analyses at the
information levels specified by I.}

\item{b}{Upper efficacy boundaries for the interim analyses at the
information levels specified by I.}

\item{sides}{\code{1=} 1-sided test (default) \cr \code{2=} 2-sided test.}

\item{r}{Integer value controlling grid for numerical integration as in
Jennison and Turnbull (2000); default is 18, range is 1 to 80.  Larger
values provide larger number of grid points and greater accuracy.  Normally
\code{r} will not be changed by the user.}
}
\value{
Object containing the following elements: \item{k}{Number of interim
 analyses.} \item{theta}{As input.} \item{I}{As input.} \item{lower}{List
 containing the lower futility boundaries (\code{bound}) and the
 corresponding probabilities of crossing the lower boundaries given
 \code{theta} (\code{prob}).} \item{upper}{List containing the upper efficacy
 boundaries (\code{bound}) and the corresponding probabilities of crossing
 the upper boundaries given theta (\code{prob}).} \item{power}{Estimated
 power for trial.} \item{futile}{Estimated probability of futility for trial}
 \item{r}{As input.}
}
\description{
Compute stopping boundary crossing probabilities.
}
\references{
Jennison C and Turnbull BW (2000), \emph{Group Sequential
 Methods with Applications to Clinical Trials}. Boca Raton: Chapman and Hall.
}
\author{
Lee Ding \email{lee_ding@g.harvard.edu}
}
