% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_sim_crt.R
\name{gs_sim_crt}
\alias{gs_sim_crt}
\title{Simulate group sequential cluster-randomized trial.}
\usage{
gs_sim_crt(design, data, stat = c("Z_known", "Z_unknown", "t_unknown"))
}
\arguments{
\item{design}{Object of class \code{gs_design_crt} specifying the group
sequential cluster-randomized trial design.}

\item{data}{Simulated outcomes. Should be an n x 4 matrix with the
columns encoding the treatment arm, cluster, individual, and response.}

\item{stat}{Type of test statistic to use. Options are \code{"Z_known"},
\code{"Z_unknown"}, and \code{"t_unknown"}.}
}
\value{
Object containing the following elements: \item{reject}{Whether the
null hypothesis was rejected in the simulated trial.} \item{k_i}{Interim
analysis at which simulated trial was stopped.} \item{m_i}{Number of clusters
per arm when the simulated trial was stopped.} \item{n_i}{Average number of
individuals per cluster when the simulated trial was stopped.}
\item{total_i}{Total number of individuals per arm when the simulated trial
was stopped.} \item{i_frac}{Information fraction when the simulated trial
was stopped.}
}
\description{
Simulate group sequential cluster-randomized trial.
}
\author{
Lee Ding \email{lee_ding@g.harvard.edu}
}
