% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsSpending.R
\name{sf_logistic}
\alias{sf_logistic}
\alias{sf_beta_dist}
\alias{sf_cauchy}
\alias{sf_extreme_value}
\alias{sf_extreme_value2}
\alias{sf_normal}
\title{Two-parameter Spending Function Families}
\usage{
sf_logistic(alpha, t, param)

sf_beta_dist(alpha, t, param)

sf_cauchy(alpha, t, param)

sf_extreme_value(alpha, t, param)

sf_extreme_value2(alpha, t, param)

sf_normal(alpha, t, param)
}
\arguments{
\item{alpha}{Real value \eqn{> 0} and no more than 1. Normally,
\code{alpha=0.025} for one-sided Type I error specification or
\code{alpha=0.1} for Type II error specification. However, this could be set
to 1 if for descriptive purposes you wish to see the proportion of spending
as a function of the proportion of sample size or information.}

\item{t}{A vector of points with increasing values from 0 to 1, inclusive.
Values of the proportion of sample size or information for which the
spending function will be computed.}

\item{param}{In the two-parameter specification, \code{sf_beta_dist()}
requires 2 positive values, while \code{sf_logistic()}, \code{sf_normal()},
\code{sf_extreme_value()},

\code{sf_extreme_value2()} and \code{sf_cauchy()} require the first parameter
to be any real value and the second to be a positive value.  The four
parameter specification is \code{c(t1,t2,u1,u2)} where the objective is that
\code{sf(t1)=alpha*u1} and \code{sf(t2)=alpha*u2}.  In this
parameterization, all four values must be between 0 and 1 and \code{t1 <
t2}, \code{u1 < u2}.}
}
\value{
An object of type \code{spendfn}. See
\code{vignette("spending_function_overview")} for further details.
}
\description{
The functions \code{sf_logistic()}, \code{sf_normal()},
\code{sf_extreme_value()}, \code{sf_extreme_value2()}, \code{sf_cauchy()}, and
\code{sf_beta_dist()} are all 2-parameter spending function families. These
provide increased flexibility in some situations where the flexibility of a
one-parameter spending function family is not sufficient. These functions
all allow fitting of two points on a cumulative spending function curve; in
this case, four parameters are specified indicating an x and a y coordinate
for each of 2 points.

\code{sf_beta_dist(alpha,t,param)} is simply \code{alpha} times the incomplete
beta cumulative distribution function with parameters \eqn{a} and \eqn{b}
passed in \code{param} evaluated at values passed in \code{t}.

The other spending functions take the form \deqn{f(t;\alpha,a,b)=\alpha
F(a+b_f^{-1}(t))}{f(t;alpha,a,b)=alpha F(a+b_f^{-1}(t))} where \eqn{F()} is a
cumulative distribution function with values \eqn{> 0} on the real line
(logistic for \code{sf_logistic()}, normal for \code{sf_normal()}, extreme
value for \code{sf_extreme_value()} and Cauchy for \code{sf_cauchy()}) and
\eqn{F^{-1}()} is its inverse.

For the logistic spending function this simplifies to
\deqn{f(t;\alpha,a,b)=\alpha (1-(1+e^a(t/(1-t))^b)^{-1}).}

For the extreme value distribution with \deqn{F(x)=\exp(-\exp(-x))} this
simplifies to \deqn{f(t;\alpha,a,b)=\alpha \exp(-e^a (-\ln t)^b).} Since the
extreme value distribution is not symmetric, there is also a version where
the standard distribution is flipped about 0. This is reflected in
\code{sf_extreme_value2()} where \deqn{F(x)=1-\exp(-\exp(x)).}
}
\note{
The gsDesign technical manual is available at
  \url{https://keaven.github.io/gsd-tech-manual/}.
}
\references{
Jennison C and Turnbull BW (2000), \emph{Group Sequential
Methods with Applications to Clinical Trials}. Boca Raton: Chapman and Hall.
}
\seealso{
\code{vignette("spending_function_overview")},
\code{\link{gs_design_crt}}, \code{vignette("gs_design_crt_package_overview")}
}
\author{
Keaven Anderson \email{keaven_anderson@merck.com}
}
\keyword{design}
