% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsSpending.R
\name{sf_exponential}
\alias{sf_exponential}
\title{Exponential Spending Function}
\usage{
sf_exponential(alpha, t, param)
}
\arguments{
\item{alpha}{Real value \eqn{> 0} and no more than 1. Normally,
\code{alpha=0.025} for one-sided Type I error specification or
\code{alpha=0.1} for Type II error specification. However, this could be set
to 1 if for descriptive purposes you wish to see the proportion of spending
as a function of the proportion of sample size/information.}

\item{t}{A vector of points with increasing values from 0 to 1, inclusive.
Values of the proportion of sample size/information for which the spending
function will be computed.}

\item{param}{A single positive value specifying the nu parameter for which
the exponential spending is to be computed; allowable range is (0, 1.5].}
}
\value{
An object of type \code{spendfn}. See
\code{vignette("spending_function_overview")} for further details.
}
\description{
The function \code{sf_exponential} implements the exponential
spending function (Anderson and Clark, 2009).

An exponential spending function is defined for any positive \code{nu} and
\eqn{0\le t\le 1} as
\deqn{f(t;\alpha,\nu)=\alpha(t)=\alpha^{t^{-\nu}}.}{f(t;alpha,nu)=alpha^(t^(-nu)).}
A value of \code{nu=0.8} approximates an O'Brien-Fleming spending function
well.

The general class of spending functions this family is derived from requires
a continuously increasing cumulative distribution function defined for
\eqn{x>0} and is defined as \deqn{f(t;\alpha,
\nu)=1-F\left(F^{-1}(1-\alpha)/ t^\nu\right).}{% f(t; alpha,
nu)=1-F(F^(-1)(1-alpha)/ t^nu).} The exponential spending function can be
derived by letting \eqn{F(x)=1-\exp(-x)}, the exponential cumulative
distribution function. This function was derived as a generalization of the
Lan-DeMets (1983) spending function used to approximate an O'Brien-Fleming
spending function (\code{sf_ldof()}), \deqn{f(t; \alpha)=2-2\Phi \left(
\Phi^{-1}(1-\alpha/2)/ t^{1/2} \right).}{% f(t;
alpha)=2-2*Phi(Phi^(-1)(1-alpha/2)/t^(1/2)).}
}
\note{
The gsDesign technical manual shows how to use \code{sf_exponential()}
  to closely approximate an O'Brien-Fleming design.
  The manual is available at <https://keaven.github.io/gsd-tech-manual/>.
}
\references{
Anderson KM and Clark JB (2009), Fitting spending functions.
\emph{Statistics in Medicine}; 29:321-327.

Jennison C and Turnbull BW (2000), \emph{Group Sequential Methods with
Applications to Clinical Trials}. Boca Raton: Chapman and Hall.

Lan, KKG and DeMets, DL (1983), Discrete sequential boundaries for clinical
trials. \emph{Biometrika}; 70:659-663.
}
\seealso{
\code{vignette("spending_function_overview")},
\code{\link{gs_design_crt}}, \code{vignette("gs_design_crt_package_overview")}
}
\author{
Keaven Anderson \email{keaven_anderson@merck.com}
}
\keyword{design}
