% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsSpending.R
\name{sf_hsd}
\alias{sf_hsd}
\title{Hwang-Shih-DeCani Spending Function}
\usage{
sf_hsd(alpha, t, param)
}
\arguments{
\item{alpha}{Real value \eqn{> 0} and no more than 1. Normally,
\code{alpha=0.025} for one-sided Type I error specification or
\code{alpha=0.1} for Type II error specification. However, this could be set
to 1 if for descriptive purposes you wish to see the proportion of spending
as a function of the proportion of sample size/information.}

\item{t}{A vector of points with increasing values from 0 to 1, inclusive.
Values of the proportion of sample size/information for which the spending
function will be computed.}

\item{param}{A single real value specifying the gamma parameter for which
Hwang-Shih-DeCani spending is to be computed; allowable range is [-40, 40]}
}
\value{
An object of type \code{spendfn}. See
\code{vignette("spending_function_overview")} for further details.
}
\description{
The function \code{sf_hsd} implements a Hwang-Shih-DeCani
spending function.

A Hwang-Shih-DeCani spending function takes the form \deqn{f(t;\alpha,
\gamma)=\alpha(1-e^{-\gamma t})/(1-e^{-\gamma})}{f(t; alpha, gamma) = alpha
* (1-exp(-gamma * t))/(1 - exp(-gamma))} where \eqn{\gamma}{gamma} is the
value passed in \code{param}. A value of \eqn{\gamma=-4}{gamma=-4} is used
to approximate an O'Brien-Fleming design (see \code{\link{sf_exponential}}
for a better fit), while a value of \eqn{\gamma=1}{gamma=1} approximates a
Pocock design well.
}
\note{
The gsDesign technical manual is available at
  \url{https://keaven.github.io/gsd-tech-manual/}.
}
\references{
Jennison C and Turnbull BW (2000), \emph{Group Sequential
Methods with Applications to Clinical Trials}. Boca Raton: Chapman and Hall.
}
\seealso{
\code{vignette("spending_function_overview")},
\code{\link{gs_design_crt}}, \code{vignette("gs_design_crt_package_overview")}
}
\author{
Keaven Anderson \email{keaven_anderson@merck.com}
}
\keyword{design}
