% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsSpending.R
\name{sf_ldof}
\alias{sf_ldof}
\alias{sf_ld_pocock}
\title{Lan-DeMets Spending function overview}
\usage{
sf_ldof(alpha, t, param = NULL)

sf_ld_pocock(alpha, t, param)
}
\arguments{
\item{alpha}{Real value \eqn{> 0} and no more than 1. Normally,
\code{alpha=0.025} for one-sided Type I error specification or
\code{alpha=0.1} for Type II error specification. However, this could be set
to 1 if for descriptive purposes you wish to see the proportion of spending
as a function of the proportion of sample size/information.}

\item{t}{A vector of points with increasing values from 0 to 1, inclusive.
Values of the proportion of sample size/information for which the spending
function will be computed.}

\item{param}{This parameter is not used for \code{sf_ld_pocock}, not required
for \code{sf_ldof} and need not be specified. For \code{sf_ld_pocock} it is here
so that the calling sequence conforms to the standard for spending functions
used with \code{gs_design_crt()}. For \code{sf_ldof} it will default to 1
(Lan-DeMets function to approximate O'Brien-Fleming) if \code{NULL} or if
outside of the range \code{[.005,2]}. otherwise, it will be use to set rho
from Liu et al (2012).}
}
\value{
An object of type \code{spendfn}. See
\code{vignette("spending_function_overview")} for further details.
}
\description{
Lan and DeMets (1983) first published the method of using
spending functions to set boundaries for group sequential trials. In this
publication they proposed two specific spending functions: one to approximate
an O'Brien-Fleming design and the other to approximate a Pocock design.
The spending function to approximate O'Brien-Fleming has been generalized as
proposed by Liu, et al (2012)

With \code{param=1=rho}, the Lan-DeMets (1983) spending function to
approximate an O'Brien-Fleming bound is implemented in the function
(\code{sf_ldof()}): \deqn{f(t; 
\alpha)=2-2\Phi\left(\Phi^{-1}(1-\alpha/2)/ t^{\rho/2}\right).}{%
f(t; alpha)=2-2*Phi(Phi^(-1)(1-alpha/2)/t^(rho/2)\right)}
For \code{rho} otherwise in \code{[.005,2]}, this is the generalized version
of Liu et al (2012). For \code{param} outside of \code{[.005,2]}, \code{rho}
is set to 1. The Lan-DeMets (1983) spending function to approximate a Pocock
design is implemented in the function \code{sf_ld_pocock()}:
\deqn{f(t;\alpha)=\alpha ln(1+(e-1)t).}{f(t;alpha)= alpha ln(1+(e-1)t).} As
shown in examples below, other spending functions can be used to ge t as good
or better approximations to Pocock and O'Brien-Fleming bounds. In particular,
O'Brien-Fleming bounds can be closely approximated using
\code{\link{sf_exponential}}.
}
\note{
The gsDesign technical manual is available at
  \url{https://keaven.github.io/gsd-tech-manual/}.
}
\references{
Jennison C and Turnbull BW (2000), \emph{Group Sequential
Methods with Applications to Clinical Trials}. Boca Raton: Chapman and Hall.

Lan, KKG and DeMets, DL (1983), Discrete sequential boundaries for clinical
trials. \emph{Biometrika};70: 659-663.

Liu, Q, Lim, P, Nuamah, I, and Li, Y (2012), On adaptive error spending
approach for group sequential trials with random information levels.
\emph{Journal of biopharmaceutical statistics}; 22(4), 687-699.
}
\seealso{
\code{vignette("spending_function_overview")},
\code{\link{gs_design_crt}}, \code{vignette("gs_design_package_overview")}
}
\author{
Keaven Anderson \email{keaven_anderson@merck.com}
}
\keyword{design}
