% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsSpending.R
\name{sf_power}
\alias{sf_power}
\title{Kim-DeMets (power) Spending Function}
\usage{
sf_power(alpha, t, param)
}
\arguments{
\item{alpha}{Real value \eqn{> 0} and no more than 1. Normally,
\code{alpha=0.025} for one-sided Type I error specification or
\code{alpha=0.1} for Type II error specification. However, this could be set
to 1 if for descriptive purposes you wish to see the proportion of spending
as a function of the proportion of sample size/information.}

\item{t}{A vector of points with increasing values from 0 to 1, inclusive.
Values of the proportion of sample size/information for which the spending
function will be computed.}

\item{param}{A single, positive value specifying the \eqn{\rho}{rho}
parameter for which Kim-DeMets spending is to be computed; allowable range
is (0,50]}
}
\value{
An object of type \code{spendfn}.
See \code{vignette("spending_function_overview")} for further details.
}
\description{
The function \code{sf_power()} implements a Kim-DeMets (power)
spending function. This is a flexible, one-parameter spending function
recommended by Jennison and Turnbull (2000).

A Kim-DeMets spending function takes the form \deqn{f(t;\alpha,\rho)=\alpha
t^\rho}{f(t; alpha, rho) = alpha t^rho} where \eqn{\rho}{rho} is the value
passed in \code{param}. See examples below for a range of values of
\eqn{\rho}{rho} that may be of interest (\code{param=0.75} to \code{3} are
documented there).
}
\note{
The gsDesign technical manual is available at
  \url{https://keaven.github.io/gsd-tech-manual/}.
}
\references{
Jennison C and Turnbull BW (2000), \emph{Group Sequential
Methods with Applications to Clinical Trials}. Boca Raton: Chapman and Hall.
}
\seealso{
\code{vignette("spending_function_overview")},
\code{\link{gs_design_crt}}, \code{vignette("gs_design_crt_package_overview")}
}
\author{
Keaven Anderson \email{keaven_anderson@merck.com}
}
\keyword{design}
